+++ modules/metastorage/src/main/java/org/apache/ignite/internal/metastorage/server/persistence/RocksDbKeyValueStorage.java
@@ -23,7 +23,7 @@ import static org.apache.ignite.internal.metastorage.server.KeyValueStorageUtils
 import static org.apache.ignite.internal.metastorage.server.KeyValueStorageUtils.assertCompactionRevisionLessThanCurrent;
 import static org.apache.ignite.internal.metastorage.server.KeyValueStorageUtils.assertRequestedRevisionLessThanOrEqualToCurrent;
 import static org.apache.ignite.internal.metastorage.server.KeyValueStorageUtils.indexToCompact;
-import static org.apache.ignite.internal.metastorage.server.KeyValueStorageUtils.isLastIndex;
+import static org.apache.ignite.internal.metastorage.server.KeyValueStorageUtils.maxRevisionIndex;
 import static org.apache.ignite.internal.metastorage.server.KeyValueStorageUtils.toUtf8String;
 import static org.apache.ignite.internal.metastorage.server.Value.TOMBSTONE;
 import static org.apache.ignite.internal.metastorage.server.persistence.RocksStorageUtils.appendLong;
@@ -43,7 +43,6 @@ import static org.apache.ignite.internal.metastorage.server.persistence.StorageC
 import static org.apache.ignite.internal.metastorage.server.persistence.StorageColumnFamilyType.REVISION_TO_TS;
 import static org.apache.ignite.internal.metastorage.server.persistence.StorageColumnFamilyType.TS_TO_REVISION;
 import static org.apache.ignite.internal.metastorage.server.raft.MetaStorageWriteHandler.IDEMPOTENT_COMMAND_PREFIX;
-import static org.apache.ignite.internal.rocksdb.RocksUtils.incrementPrefix;
 import static org.apache.ignite.internal.rocksdb.snapshot.ColumnFamilyRange.fullRange;
 import static org.apache.ignite.internal.util.ArrayUtils.LONG_EMPTY_ARRAY;
 import static org.apache.ignite.internal.util.ByteUtils.toByteArray;
@@ -60,23 +59,15 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Objects;
-import java.util.TreeSet;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.concurrent.locks.ReadWriteLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
-import java.util.function.LongConsumer;
-import java.util.function.Predicate;
 import org.apache.ignite.internal.failure.FailureManager;
 import org.apache.ignite.internal.hlc.HybridTimestamp;
 import org.apache.ignite.internal.lang.ByteArray;
@@ -84,8 +75,6 @@ import org.apache.ignite.internal.logger.IgniteLogger;
 import org.apache.ignite.internal.logger.Loggers;
 import org.apache.ignite.internal.metastorage.CommandId;
 import org.apache.ignite.internal.metastorage.Entry;
-import org.apache.ignite.internal.metastorage.RevisionUpdateListener;
-import org.apache.ignite.internal.metastorage.WatchListener;
 import org.apache.ignite.internal.metastorage.dsl.Operation;
 import org.apache.ignite.internal.metastorage.dsl.OperationType;
 import org.apache.ignite.internal.metastorage.dsl.Operations;
@@ -94,17 +83,14 @@ import org.apache.ignite.internal.metastorage.dsl.Update;
 import org.apache.ignite.internal.metastorage.exceptions.CompactedException;
 import org.apache.ignite.internal.metastorage.exceptions.MetaStorageException;
 import org.apache.ignite.internal.metastorage.impl.EntryImpl;
-import org.apache.ignite.internal.metastorage.impl.MetaStorageManagerImpl;
+import org.apache.ignite.internal.metastorage.server.AbstractKeyValueStorage;
 import org.apache.ignite.internal.metastorage.server.Condition;
 import org.apache.ignite.internal.metastorage.server.If;
 import org.apache.ignite.internal.metastorage.server.KeyValueStorage;
-import org.apache.ignite.internal.metastorage.server.KeyValueStorageUtils;
 import org.apache.ignite.internal.metastorage.server.MetastorageChecksum;
 import org.apache.ignite.internal.metastorage.server.OnRevisionAppliedCallback;
 import org.apache.ignite.internal.metastorage.server.Statement;
 import org.apache.ignite.internal.metastorage.server.Value;
-import org.apache.ignite.internal.metastorage.server.Watch;
-import org.apache.ignite.internal.metastorage.server.WatchProcessor;
 import org.apache.ignite.internal.rocksdb.ColumnFamily;
 import org.apache.ignite.internal.rocksdb.RocksIteratorAdapter;
 import org.apache.ignite.internal.rocksdb.RocksUtils;
@@ -143,7 +129,7 @@ import org.rocksdb.WriteOptions;
  * The mapping from the key to the set of the storage's revisions is stored in the "index" column family. A key represents the key of an
  * entry and the value is a {@code byte[]} that represents a {@code long[]} where every item is a revision of the storage.
  */
-public class RocksDbKeyValueStorage implements KeyValueStorage {
+public class RocksDbKeyValueStorage extends AbstractKeyValueStorage {
     private static final IgniteLogger LOG = Loggers.forClass(RocksDbKeyValueStorage.class);
 
     /** A revision to store with system entries. */
@@ -161,18 +147,12 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
             "SYSTEM_COMPACTION_REVISION_KEY".getBytes(UTF_8)
     );
 
-    /** Lexicographic order comparator. */
-    private static final Comparator<byte[]> CMP = Arrays::compareUnsigned;
-
     /** Batch size (number of keys) for storage compaction. The value is arbitrary. */
     private static final int COMPACT_BATCH_SIZE = 10;
 
     static {
         RocksDB.loadLibrary();
     }
-
-    /** RW lock. */
-    private final ReadWriteLock rwLock = new ReentrantReadWriteLock();
 
     /** Thread-pool for snapshot operations execution. */
     private final ExecutorService snapshotExecutor;
@@ -205,36 +185,11 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
     private volatile RocksSnapshotManager snapshotManager;
 
     /**
-     * Revision listener for recovery only. Notifies {@link MetaStorageManagerImpl} of revision update.
-     * Guarded by {@link #rwLock}.
-     */
-    private @Nullable LongConsumer recoveryRevisionListener;
-
-    /**
-     * Revision. Will be incremented for each single-entry or multi-entry update operation.
-     *
-     * <p>Multi-threaded access is guarded by {@link #rwLock}.</p>
-     */
-    private long rev;
-
-    /**
      * Facility to work with checksums.
      *
      * <p>Multi-threaded access is guarded by {@link #rwLock}.</p>
      */
     private MetastorageChecksum checksum;
-
-    /**
-     * Last compaction revision that was set or restored from a snapshot.
-     *
-     * <p>This field is used by metastorage read methods to determine whether {@link CompactedException} should be thrown.</p>
-     *
-     * <p>Multi-threaded access is guarded by {@link #rwLock}.</p>
-     */
-    private long compactionRevision = -1;
-
-    /** Watch processor. */
-    private final WatchProcessor watchProcessor;
 
     /** Status of the watch recovery process. */
     private enum RecoveryStatus {
@@ -271,18 +226,17 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
     /** Metastorage recovery is based on the snapshot & external log. WAL is never used for recovery, and can be safely disabled. */
     private final WriteOptions defaultWriteOptions = new WriteOptions().setDisableWAL(true);
 
-    private final AtomicBoolean stopCompaction = new AtomicBoolean();
-
     /**
      * Constructor.
      *
      * @param nodeName Node name.
      * @param dbPath RocksDB path.
+     * @param failureManager Failure processor that is used to handle critical errors.
      */
     public RocksDbKeyValueStorage(String nodeName, Path dbPath, FailureManager failureManager) {
-        this.dbPath = dbPath;
+        super(nodeName, failureManager);
 
-        this.watchProcessor = new WatchProcessor(nodeName, this::get, failureManager);
+        this.dbPath = dbPath;
 
         this.snapshotExecutor = Executors.newFixedThreadPool(2, NamedThreadFactory.create(nodeName, "metastorage-snapshot-executor", LOG));
     }
@@ -411,17 +365,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
     }
 
     /**
-     * Notifies of revision update.
-     * Must be called under the {@link #rwLock}.
-     */
-    private void notifyRevisionUpdate() {
-        if (recoveryRevisionListener != null) {
-            // Listener must be invoked only on recovery, after recovery listener must be null.
-            recoveryRevisionListener.accept(rev);
-        }
-    }
-
-    /**
      * Clear the RocksDB instance.
      *
      * @throws IOException If failed.
@@ -457,7 +400,13 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
 
     @Override
     public CompletableFuture<Void> snapshot(Path snapshotPath) {
-        return snapshotManager.createSnapshot(snapshotPath);
+        rwLock.writeLock().lock();
+
+        try {
+            return snapshotManager.createSnapshot(snapshotPath);
+        } finally {
+            rwLock.writeLock().unlock();
+        }
     }
 
     @Override
@@ -487,17 +436,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
             throw new MetaStorageException(RESTORING_STORAGE_ERR, "Failed to restore snapshot", e);
         } finally {
             rwLock.writeLock().unlock();
-        }
-    }
-
-    @Override
-    public long revision() {
-        rwLock.readLock().lock();
-
-        try {
-            return rev;
-        } finally {
-            rwLock.readLock().unlock();
         }
     }
 
@@ -597,12 +535,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
         return longToBytes(ts.longValue());
     }
 
-    private static Entry entry(byte[] key, long revision, Value value) {
-        return value.tombstone()
-                ? EntryImpl.tombstone(key, revision, value.operationTimestamp())
-                : new EntryImpl(key, value.bytes(), revision, value.operationTimestamp());
-    }
-
     @Override
     public void putAll(List<byte[]> keys, List<byte[]> values, HybridTimestamp opTs) {
         rwLock.writeLock().lock();
@@ -623,61 +555,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
             throw new MetaStorageException(OP_EXECUTION_ERR, e);
         } finally {
             rwLock.writeLock().unlock();
-        }
-    }
-
-    @Override
-    public Entry get(byte[] key) {
-        rwLock.readLock().lock();
-
-        try {
-            return doGet(key, rev);
-        } finally {
-            rwLock.readLock().unlock();
-        }
-    }
-
-    @Override
-    public Entry get(byte[] key, long revUpperBound) {
-        rwLock.readLock().lock();
-
-        try {
-            return doGet(key, revUpperBound);
-        } finally {
-            rwLock.readLock().unlock();
-        }
-    }
-
-    @Override
-    public List<Entry> get(byte[] key, long revLowerBound, long revUpperBound) {
-        rwLock.readLock().lock();
-
-        try {
-            return doGet(key, revLowerBound, revUpperBound);
-        } finally {
-            rwLock.readLock().unlock();
-        }
-    }
-
-    @Override
-    public List<Entry> getAll(List<byte[]> keys) {
-        rwLock.readLock().lock();
-
-        try {
-            return doGetAll(keys, rev);
-        } finally {
-            rwLock.readLock().unlock();
-        }
-    }
-
-    @Override
-    public List<Entry> getAll(List<byte[]> keys, long revUpperBound) {
-        rwLock.readLock().lock();
-
-        try {
-            return doGetAll(keys, revUpperBound);
-        } finally {
-            rwLock.readLock().unlock();
         }
     }
 
@@ -861,17 +738,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
     }
 
     @Override
-    public Cursor<Entry> range(byte[] keyFrom, byte @Nullable [] keyTo) {
-        rwLock.readLock().lock();
-
-        try {
-            return range(keyFrom, keyTo, rev);
-        } finally {
-            rwLock.readLock().unlock();
-        }
-    }
-
-    @Override
     public Cursor<Entry> range(byte[] keyFrom, byte @Nullable [] keyTo, long revUpperBound) {
         rwLock.readLock().lock();
 
@@ -931,9 +797,9 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
                 protected Entry decodeEntry(byte[] key, byte[] value) {
                     long[] revisions = getAsLongs(value);
 
-                    long targetRevision = maxRevision(revisions, revUpperBound);
+                    long targetRevision = maxRevisionIndex(revisions, revUpperBound);
 
-                    if (targetRevision == -1) {
+                    if (targetRevision == NOT_FOUND) {
                         return EntryImpl.empty(key);
                     }
 
@@ -962,44 +828,8 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
     }
 
     @Override
-    public void watchRange(byte[] keyFrom, byte @Nullable [] keyTo, long rev, WatchListener listener) {
-        assert keyFrom != null : "keyFrom couldn't be null.";
-        assert rev > 0 : "rev must be positive.";
-
-        Predicate<byte[]> rangePredicate = keyTo == null
-                ? k -> CMP.compare(keyFrom, k) <= 0
-                : k -> CMP.compare(keyFrom, k) <= 0 && CMP.compare(keyTo, k) > 0;
-
-        watchProcessor.addWatch(new Watch(rev, listener, rangePredicate));
-    }
-
-    @Override
-    public void watchExact(byte[] key, long rev, WatchListener listener) {
-        assert key != null : "key couldn't be null.";
-        assert rev > 0 : "rev must be positive.";
-
-        Predicate<byte[]> exactPredicate = k -> CMP.compare(k, key) == 0;
-
-        watchProcessor.addWatch(new Watch(rev, listener, exactPredicate));
-    }
-
-    @Override
-    public void watchExact(Collection<byte[]> keys, long rev, WatchListener listener) {
-        assert keys != null && !keys.isEmpty() : "keys couldn't be null or empty: " + keys;
-        assert rev > 0 : "rev must be positive.";
-
-        TreeSet<byte[]> keySet = new TreeSet<>(CMP);
-
-        keySet.addAll(keys);
-
-        Predicate<byte[]> inPredicate = keySet::contains;
-
-        watchProcessor.addWatch(new Watch(rev, listener, inPredicate));
-    }
-
-    @Override
     public void startWatches(long startRevision, OnRevisionAppliedCallback revisionCallback) {
-        assert startRevision != 0 : "First meaningful revision is 1";
+        assert startRevision > 0 : startRevision;
 
         long currentRevision;
 
@@ -1028,11 +858,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
     }
 
     @Override
-    public void removeWatch(WatchListener listener) {
-        watchProcessor.removeWatch(listener);
-    }
-
-    @Override
     public void compact(long revision) {
         assert revision >= 0 : revision;
 
@@ -1043,16 +868,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
         } catch (Throwable t) {
             throw new MetaStorageException(COMPACTION_ERR, "Error during compaction: " + revision, t);
         }
-    }
-
-    @Override
-    public void stopCompaction() {
-        stopCompaction.set(true);
-    }
-
-    @Override
-    public byte @Nullable [] nextKey(byte[] key) {
-        return incrementPrefix(key);
     }
 
     /**
@@ -1121,87 +936,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
         }
     }
 
-    private List<Entry> doGetAll(Collection<byte[]> keys, long revUpperBound) {
-        assert !keys.isEmpty();
-        assert revUpperBound >= 0 : revUpperBound;
-
-        var res = new ArrayList<Entry>(keys.size());
-
-        for (byte[] key : keys) {
-            res.add(doGet(key, revUpperBound));
-        }
-
-        return res;
-    }
-
-    private Entry doGet(byte[] key, long revUpperBound) {
-        assert revUpperBound >= 0 : revUpperBound;
-
-        long[] keyRevisions = getRevisionsForOperation(key);
-        int maxRevisionIndex = KeyValueStorageUtils.maxRevisionIndex(keyRevisions, revUpperBound);
-
-        if (maxRevisionIndex == NOT_FOUND) {
-            CompactedException.throwIfRequestedRevisionLessThanOrEqualToCompacted(revUpperBound, compactionRevision);
-
-            return EntryImpl.empty(key);
-        }
-
-        long revision = keyRevisions[maxRevisionIndex];
-
-        Value value = getValueForOperation(key, revision);
-
-        if (revUpperBound <= compactionRevision && (!isLastIndex(keyRevisions, maxRevisionIndex) || value.tombstone())) {
-            throw new CompactedException(revUpperBound, compactionRevision);
-        }
-
-        return EntryImpl.toEntry(key, revision, value);
-    }
-
-    /**
-     * Returns all entries corresponding to the given key and bounded by given revisions.
-     * All these entries are ordered by revisions and have the same key.
-     * The lower bound and the upper bound are inclusive.
-     *
-     * @param key The key.
-     * @param revLowerBound The lower bound of revision.
-     * @param revUpperBound The upper bound of revision.
-     * @return Entries corresponding to the given key.
-     */
-    private List<Entry> doGet(byte[] key, long revLowerBound, long revUpperBound) {
-        assert revLowerBound >= 0 : "Invalid arguments: [revLowerBound=" + revLowerBound + ']';
-        assert revUpperBound >= 0 : "Invalid arguments: [revUpperBound=" + revUpperBound + ']';
-        assert revUpperBound >= revLowerBound
-                : "Invalid arguments: [revLowerBound=" + revLowerBound + ", revUpperBound=" + revUpperBound + ']';
-
-        long[] revs;
-
-        try {
-            revs = getRevisions(key);
-        } catch (RocksDBException e) {
-            throw new MetaStorageException(OP_EXECUTION_ERR, e);
-        }
-
-        if (revs.length == 0) {
-            return Collections.emptyList();
-        }
-
-        int firstRevIndex = minRevisionIndex(revs, revLowerBound);
-        int lastRevIndex = maxRevisionIndex(revs, revUpperBound);
-
-        // firstRevIndex can be -1 if minRevisionIndex return -1. lastRevIndex can be -1 if maxRevisionIndex return -1.
-        if (firstRevIndex == -1 || lastRevIndex == -1) {
-            return Collections.emptyList();
-        }
-
-        List<Entry> entries = new ArrayList<>();
-
-        for (int i = firstRevIndex; i <= lastRevIndex; i++) {
-            entries.add(doGetValue(key, revs[i]));
-        }
-
-        return entries;
-    }
-
     /**
      * Returns array of revisions of the entry corresponding to the key.
      *
@@ -1216,80 +950,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
         }
 
         return getAsLongs(revisions);
-    }
-
-    /**
-     * Returns array of revisions of the entry corresponding to the key.
-     *
-     * @param key Key.
-     * @throws MetaStorageException If there was an error while getting the revisions for the key.
-     */
-    private long[] getRevisionsForOperation(byte[] key) {
-        try {
-            return getRevisions(key);
-        } catch (RocksDBException e) {
-            throw new MetaStorageException(OP_EXECUTION_ERR, "Failed to get revisions for the key: " + toUtf8String(key), e);
-        }
-    }
-
-    /**
-     * Returns maximum revision which must be less or equal to {@code upperBoundRev}. If there is no such revision then {@code -1} will be
-     * returned.
-     *
-     * @param revs          Revisions list.
-     * @param upperBoundRev Revision upper bound.
-     * @return Maximum revision or {@code -1} if there is no such revision.
-     */
-    private static long maxRevision(long[] revs, long upperBoundRev) {
-        for (int i = revs.length - 1; i >= 0; i--) {
-            long rev = revs[i];
-
-            if (rev <= upperBoundRev) {
-                return rev;
-            }
-        }
-
-        return -1;
-    }
-
-    /**
-     * Returns index of minimum revision which must be greater or equal to {@code lowerBoundRev}.
-     * If there is no such revision then {@code -1} will be returned.
-     *
-     * @param revs          Revisions list.
-     * @param lowerBoundRev Revision lower bound.
-     * @return Index of minimum revision or {@code -1} if there is no such revision.
-     */
-    private static int minRevisionIndex(long[] revs, long lowerBoundRev) {
-        for (int i = 0; i < revs.length; i++) {
-            long rev = revs[i];
-
-            if (rev >= lowerBoundRev) {
-                return i;
-            }
-        }
-
-        return -1;
-    }
-
-    /**
-     * Returns index of maximum revision which must be less or equal to {@code upperBoundRev}.
-     * If there is no such revision then {@code -1} will be returned.
-     *
-     * @param revs          Revisions list.
-     * @param upperBoundRev Revision upper bound.
-     * @return Index of maximum revision or {@code -1} if there is no such revision.
-     */
-    private static int maxRevisionIndex(long[] revs, long upperBoundRev) {
-        for (int i = revs.length - 1; i >= 0; i--) {
-            long rev = revs[i];
-
-            if (rev <= upperBoundRev) {
-                return i;
-            }
-        }
-
-        return -1;
     }
 
     /**
@@ -1348,7 +1008,7 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
 
         data.put(batch, rocksKey, rocksValue);
 
-        updatedEntries.add(entry(key, curRev, new Value(value, opTs)));
+        updatedEntries.add(EntryImpl.toEntry(key, curRev, new Value(value, opTs)));
     }
 
     /**
@@ -1464,7 +1124,7 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
                     ts = hybridTimestamp(timestampFromRocksValue(rocksValue));
                 }
 
-                updatedEntries.add(entry(rocksKeyToBytes(rocksKey), revision, bytesToValue(rocksValue)));
+                updatedEntries.add(EntryImpl.toEntry(rocksKeyToBytes(rocksKey), revision, bytesToValue(rocksValue)));
             }
 
             try {
@@ -1551,17 +1211,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
         }
     }
 
-    @Override
-    public void setRecoveryRevisionListener(@Nullable LongConsumer listener) {
-        rwLock.writeLock().lock();
-
-        try {
-            this.recoveryRevisionListener = listener;
-        } finally {
-            rwLock.writeLock().unlock();
-        }
-    }
-
     @TestOnly
     public Path getDbPath() {
         return dbPath;
@@ -1608,21 +1257,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
     }
 
     @Override
-    public void registerRevisionUpdateListener(RevisionUpdateListener listener) {
-        watchProcessor.registerRevisionUpdateListener(listener);
-    }
-
-    @Override
-    public void unregisterRevisionUpdateListener(RevisionUpdateListener listener) {
-        watchProcessor.unregisterRevisionUpdateListener(listener);
-    }
-
-    @Override
-    public CompletableFuture<Void> notifyRevisionUpdateListenerOnStart(long newRevision) {
-        return watchProcessor.notifyUpdateRevisionListeners(newRevision);
-    }
-
-    @Override
     public void advanceSafeTime(HybridTimestamp newSafeTime) {
         rwLock.writeLock().lock();
 
@@ -1651,32 +1285,6 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
             throw new MetaStorageException(COMPACTION_ERR, "Error saving compaction revision: " + revision, t);
         } finally {
             rwLock.writeLock().unlock();
-        }
-    }
-
-    @Override
-    public void setCompactionRevision(long revision) {
-        assert revision >= 0 : revision;
-
-        rwLock.writeLock().lock();
-
-        try {
-            assertCompactionRevisionLessThanCurrent(revision, rev);
-
-            compactionRevision = revision;
-        } finally {
-            rwLock.writeLock().unlock();
-        }
-    }
-
-    @Override
-    public long getCompactionRevision() {
-        rwLock.readLock().lock();
-
-        try {
-            return compactionRevision;
-        } finally {
-            rwLock.readLock().unlock();
         }
     }
 
@@ -1788,7 +1396,17 @@ public class RocksDbKeyValueStorage implements KeyValueStorage {
         }
     }
 
-    private Value getValueForOperation(byte[] key, long revision) {
+    @Override
+    protected long[] keyRevisionsForOperation(byte[] key) {
+        try {
+            return getRevisions(key);
+        } catch (RocksDBException e) {
+            throw new MetaStorageException(OP_EXECUTION_ERR, "Failed to get revisions for the key: " + toUtf8String(key), e);
+        }
+    }
+
+    @Override
+    protected Value valueForOperation(byte[] key, long revision) {
         try {
             byte[] valueBytes = data.get(keyToRocksKey(revision, key));
